Added PeakWindow class allowing multiple peak windows to be open at once, made changes to removing graphs; will now call clear func if removing the only graph.
